migrate form old Nighthawk endpoints
